+++ providers/ESPN.js
@@ -420,6 +420,7 @@ module.exports = {
   broadcastIcons: {
     'ACC Network': 'https://upload.wikimedia.org/wikipedia/commons/f/fd/ACC_Network_ESPN_logo.svg',
     'ACC Extra': './modules/MMM-MyScoreboard/logos/channels/ACC-NX-logo.png',
+    'Apple TV': 'https://upload.wikimedia.org/wikipedia/en/e/eb/Apple_TV_logo_2025.png',
     'B1G+': './modules/MMM-MyScoreboard/logos/channels/B1GPlus.png',
     'BTN': 'https://upload.wikimedia.org/wikipedia/en/3/39/Big_Ten_Network_Logo.svg',
     'CW': 'https://upload.wikimedia.org/wikipedia/commons/b/b1/The_CW_2024.svg',